requirements.txt
pvporcupine
pyaudio
vosk
openai
anthropic
google-generativeai
requests
sounddevice
soundfile
numpy
scipy
pyautogui
dbus-python
playsound3
coqui-tts
openai-whisper
docker
moviepy
librosa
Pillow
beautifulsoup4
